src/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.13.0; python_version >= '3.10'
asgiref==3.11.1; python_version >= '3.9'
backports-datetime-fromisoformat==2.0.3; python_version >= '3.1'
boto3==1.42.82; python_version >= '3.9'
botocore==1.42.82; python_version >= '3.9'
cachetools==7.0.5; python_version >= '3.10'
certifi==2026.2.25; python_version >= '3.7'
cfenv==0.5.3
cffi==2.0.0; python_version >= '3.9'
charset-normalizer==3.4.7; python_version >= '3.7'
cryptography==46.0.6; python_version >= '3.8' and python_full_version not in '3.9.0, 3.9.1'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
diff-match-patch==20241021; python_version >= '3.7'
dj-database-url==3.1.2; python_version >= '3.10'
dj-email-url==1.0.6
django==4.2.29; python_version >= '3.8'
django-admin-multiple-choice-list-filter==0.1.1
django-allow-cidr==0.8.0
django-auditlog==3.4.1; python_version >= '3.10'
django-cache-url==3.4.6
django-cors-headers==4.9.0; python_version >= '3.9'
django-csp==4.0
django-fsm==2.8.1
django-htmx==1.27.0; python_version >= '3.10'
django-import-export==4.4.0; python_version >= '3.9'
django-login-required-middleware==0.9.0
django-phonenumber-field[phonenumberslite]==8.4.0; python_version >= '3.10'
django-waffle==5.0.0; python_version >= '3.9'
django-widget-tweaks==1.5.1; python_version >= '3.9'
environs[django]==15.0.0; python_version >= '3.10'
exceptiongroup==1.3.1; python_version >= '3.7'
faker==40.12.0; python_version >= '3.10'
fred-epplib @ git+https://github.com/cisagov/epplib.git@308e625b794a9a8e50551e8f9caa837470e24630
furl==2.1.4
future==1.0.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
gevent==25.9.1; python_version >= '3.9'
greenlet==3.3.2; python_version >= '3.10'
gunicorn==25.3.0; python_version >= '3.10'
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
idna==3.11; python_version >= '3.8'
jmespath==1.1.0; python_version >= '3.9'
lxml==6.0.2; python_version >= '3.8'
mako==1.3.10; python_version >= '3.8'
markupsafe==3.0.3; python_version >= '3.9'
marshmallow==4.2.4; python_version >= '3.10'
oic==1.7.0; python_version ~= '3.8'
orderedmultidict==1.0.2
packaging==26.0; python_version >= '3.8'
phonenumberslite==9.0.27; python_version >= '2.5'
pillow==12.2.0; python_version >= '3.10'
psycopg2-binary==2.9.11; python_version >= '3.9'
pycparser==3.0; python_version >= '3.10'
pycryptodomex==3.23.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
pydantic==2.12.5; python_version >= '3.9'
pydantic-core==2.41.5; python_version >= '3.9'
pydantic-settings==2.13.1; python_version >= '3.10'
pyjwkest==1.4.4
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.2.2; python_version >= '3.10'
pyzipper==0.3.6; python_version >= '3.4'
rapidfuzz==3.14.3; python_version >= '3.10'
requests==2.33.1; python_version >= '3.10'
respx==0.22.0; python_version >= '3.8'
s3transfer==0.16.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.5.5; python_version >= '3.8'
tablib==3.9.0; python_version >= '3.9'
tblib==3.2.2; python_version >= '3.9'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.2; python_version >= '3.9'
urllib3==2.6.3; python_version >= '3.9'
whitenoise==6.12.0; python_version >= '3.10'
zope.event==6.1; python_version >= '3.10'
zope.interface==8.2; python_version >= '3.10'
